Fix warning

+++ ApplicationLibCode/GeoMech/GeoMechDataModel/RigFemPartResultsCollection.cpp
@@ -1013,7 +1013,7 @@ std::vector<std::string> RigFemPartResultsCollection::stepNames() const
 const std::pair<int, int> RigFemPartResultsCollection::stepListIndexToTimeStepAndDataFrameIndex( int stepIndex ) const
 {
     if ( stepIndex < 0 ) return std::make_pair( stepIndex, -1 );
-    CVF_ASSERT( stepIndex < m_stepList.size() );
+    CVF_ASSERT( stepIndex < static_cast<int>( m_stepList.size() ) );
     return m_stepList[stepIndex];
 }
